More warning flags.

+++ src/main.c
@@ -28,7 +28,7 @@ int main(void) {
     add_new_line();
 
     // * Display correct details.
-    for (int i = 0; i < (sizeof(elements) / sizeof(element)); i++) {
+    for (long unsigned int i = 0; i < (sizeof(elements) / sizeof(element)); i++) {
       element current_element = elements[i];
       char *current_symbol = current_element.symbol;
       char *current_name = current_element.name;
